Refactor module packaging
