+++ http/http_client_jetty_ws/api/http_client_jetty_ws.api
@@ -1,4 +1,4 @@
-public final class com/hexagonkt/http/client/jetty/JettyClientWsSession : com/hexagonkt/http/model/ws/WsSession {
+public final class com/hexagonkt/http/client/jetty/ws/JettyClientWsSession : com/hexagonkt/http/model/ws/WsSession {
 	public fun <init> (Ljava/net/URI;Lorg/eclipse/jetty/websocket/api/Session;)V
 	public fun close (ILjava/lang/String;)V
 	public fun getAttributes ()Ljava/util/Map;
@@ -12,7 +12,7 @@ public final class com/hexagonkt/http/client/jetty/JettyClientWsSession : com/he
 	public fun send ([B)V
 }
 
-public final class com/hexagonkt/http/client/jetty/JettyWebSocketAdapter {
+public final class com/hexagonkt/http/client/jetty/ws/JettyWebSocketAdapter {
 	public fun <init> (Ljava/net/URI;Lkotlin/jvm/functions/Function1;Lkotlin/jvm/functions/Function2;Lkotlin/jvm/functions/Function2;Lkotlin/jvm/functions/Function2;Lkotlin/jvm/functions/Function2;Lkotlin/jvm/functions/Function3;)V
 	public final fun onWebSocketBinary (Lorg/eclipse/jetty/websocket/api/Session;Ljava/nio/ByteBuffer;Lorg/eclipse/jetty/websocket/api/Callback;)V
 	public final fun onWebSocketClose (Lorg/eclipse/jetty/websocket/api/Session;ILjava/lang/String;)V
@@ -20,7 +20,7 @@ public final class com/hexagonkt/http/client/jetty/JettyWebSocketAdapter {
 	public final fun onWebSocketText (Lorg/eclipse/jetty/websocket/api/Session;Ljava/lang/String;)V
 }
 
-public final class com/hexagonkt/http/client/jetty/JettyWsClientAdapter : com/hexagonkt/http/client/jetty/JettyClientAdapter {
+public final class com/hexagonkt/http/client/jetty/ws/JettyWsClientAdapter : com/hexagonkt/http/client/jetty/JettyClientAdapter {
 	public fun <init> ()V
 	public fun shutDown ()V
 	public fun startUp (Lcom/hexagonkt/http/client/HttpClient;)V
