+++ http/http_client_jetty_ws/src/main/kotlin/com/hexagonkt/http/client/jetty/ws/JettyWebSocketAdapter.kt
@@ -1,4 +1,4 @@
-package com.hexagonkt.http.client.jetty
+package com.hexagonkt.http.client.jetty.ws
 
 import com.hexagonkt.http.model.ws.WsSession
 import org.eclipse.jetty.websocket.api.Callback
