+++ http/http_client_jetty_ws/src/main/kotlin/com/hexagonkt/http/client/jetty/ws/JettyWsClientAdapter.kt
@@ -1,7 +1,8 @@
-package com.hexagonkt.http.client.jetty
+package com.hexagonkt.http.client.jetty.ws
 
 import com.hexagonkt.core.urlOf
 import com.hexagonkt.http.client.HttpClient
+import com.hexagonkt.http.client.jetty.JettyClientAdapter
 import com.hexagonkt.http.model.ws.WsSession
 import org.eclipse.jetty.websocket.client.WebSocketClient
 import java.net.URI
