+++ http/http_server_jetty/src/test/kotlin/com/hexagonkt/http/server/jetty/JettyServletAdapterTest.kt
@@ -4,7 +4,7 @@ import com.hexagonkt.core.media.TEXT_PLAIN
 import com.hexagonkt.core.urlOf
 import com.hexagonkt.http.client.HttpClient
 import com.hexagonkt.http.client.HttpClientSettings
-import com.hexagonkt.http.client.jetty.JettyWsClientAdapter
+import com.hexagonkt.http.client.jetty.ws.JettyWsClientAdapter
 import com.hexagonkt.http.model.ContentType
 import com.hexagonkt.http.model.OK_200
 import com.hexagonkt.http.handlers.path
